rename service files, add createProduct
