+++ react-cors-spa/src/services/CategoryServices.js
