+++ react-cors-spa/src/services/ProductServices.js
@@ -4,4 +4,8 @@ function fetchProducts() {
     return axios.get('http://127.0.0.1:3000/products?limit=5&offset=0');
 }
 
-export default fetchProducts;
+function createProduct() {
+    return axios.get('http://127.0.0.1:3000/products')
+}
+
+export { fetchProducts, createProduct };
